Change cdjot_convert len parameter from int to size_t
Use size_t for the length parameter in the public API, which is more
idiomatic for a byte count in C and avoids the 2GB int limit.

+++ cdjot.c
@@ -2873,7 +2873,7 @@ emit_endnotes(void)
 }
 
 int
-cdjot_convert(FILE *out, const char *buf, int len)
+cdjot_convert(FILE *out, const char *buf, size_t len)
 {
 	int i;
